Crate pallet_defense

Defense Pallet

A secure module for protecting users’ asset

Overview

The Defense module provides functionality for asset protection of asset

Including:

set_transfer_limit

set_risk_management

safe_transfer

Re-exports

pub use pallet::*;

Modules

crypto

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Enums

RiskManagement

a enum to organize different protect actions when illegal transactions occur there are two kind of predefined actions

TransferLimit

a enum to organize different limitations of transferring money there are two kind of predefined limitations

Constants

KEY_TYPE

Type Definitions

CallHash

OpaqueCall